preprocessing before training

options for preprocessing:

dummy variable encoding

modification to the whole dataset before training, including

imputation using missForest and KNN

detection and removal of Near Zero Variance (NZV) and ConditionalX variables Note that these modifications apply only in situations where caret::train is not used in the model building; for example

ML method is glm where no cross validation (CV) is needed due to the absence of hyper-parameters

ML method is randomForest where the training is done using randomForest::tuneRF() Otherwise, those modification should be specified in caret::train(preProc) to avoid information leakage during training. Note that missForest is not available in preProc.

Usage

df.preprocess(
  allDF = NULL,
  trainDF = NULL,
  testDF = NULL,
  prior.nzvCondX = c("n", "nx"),
  dummyEncoding = FALSE,
  prior.impute = c(NA, "missForest", "KNN"),
  classCol,
  posLabel = NULL,
  negLabel = NULL,
  oDir = NULL
)

Arguments

allDF

NULL, data.frame; the data.frame of all the data

trainDF

NULL, data.frame; the data.frame of the training data

testDF

NULL, data.frame; the data.frame of the testing data

prior.nzvCondX

NULL, char; remove NZV and conditionalX variables before training. This can be used only if caret::train is not used in the model building. Following are the available options: * NULL: do not remove before training * one of n, nx: + n: NZV + nx: NZV and ConditionalX if detected, removed variables are written to files nzv.csv and conditionalX.txt

dummyEncoding

cMethod %in% c('xgbTree','svmRadial'),logical; is dummy variable encoding needed before training;

prior.impute

c(NA,'missForest','KNN'), char; impute missing values before training. This can be used only if caret::train is not used in the model building. Following are the available options: * NA: do not impute before training * one of missForest, KNN: impute using missForest::missForest() or DMwR2::knnImputation()

classCol

char; the column name of the class, i.e. response variable

posLabel, negLabel

NULL, char; the positive and negative labels for two-class classification

oDir

NULL, char; output directory

Value

a list containing trainDF, testDF, and allDF post processing.

Examples

# none