convert categorical columns into numerical ones of full rank

It does the following:

for nominal character variable of n categories, n-1 binary variables are created

for nominal factor variable of n levels and m categories (m<=n), * if useLevel=TRUE, n-1 binary variables are created * otherwise, m-1 binary variables are created

for ordinal factor variable of n levels and m categories, * if m equals 2, one variable assuming equal distance among levels are created * otherwise, n-1 variable assuming linear and 2 to n-1 degree of polynomial relationship among levels are created.

cells of NA are kept.

Usage

dummy.encode(df1, useLevel = FALSE, ...)

Arguments

df1

data frame; columns are variables.

useLevel

FALSE, logical; if FALSE, for factor columns, the unique values will be factor levels, instead of the actual levels present in the data.

...

passed to caret::dummyVars()

Value

a new data.frame where all categorical columns are split into dummy columns of values between 1 and 0.

Examples

k=matrix(1:10,nrow = 5)
colnames(k)=c('n1','n2')
 k=as.data.frame(k)
 k$c5=LETTERS[1:5]
 k$c2na=c('a','a',NA,'b','b')
 k$f2na=factor(k$c2na)
 k$f2naL2O=factor(k$f2na,ordered = TRUE)
 k$f2naL3=factor(k$c2na,levels = c('a','b','c'))
 k$f2naL3O=factor(k$f2naL3,ordered = TRUE)
 k$f3na=factor(c('a','a',NA,'b','c'))
 k$f3naL3O=factor(k$f3na,ordered = TRUE)
 k$f3naL4O=factor(k$f3na,levels = c('a','b','c','d'),ordered = TRUE)
 k$f3naL5O=factor(k$f3na,levels = c('a','b','c','d','e'),ordered = TRUE)
 k
#>   n1 n2 c5 c2na f2na f2naL2O f2naL3 f2naL3O f3na f3naL3O f3naL4O f3naL5O
#> 1  1  6  A    a    a       a      a       a    a       a       a       a
#> 2  2  7  B    a    a       a      a       a    a       a       a       a
#> 3  3  8  C <NA> <NA>    <NA>   <NA>    <NA> <NA>    <NA>    <NA>    <NA>
#> 4  4  9  D    b    b       b      b       b    b       b       b       b
#> 5  5 10  E    b    b       b      b       b    c       c       c       c
 dummy.encode(k)
#>   n1 n2 c5B c5C c5D c5E c2nab f2na.b  f2naL2O.L f2naL3.b  f2naL3O.L f3na.b
#> 1  1  6   0   0   0   0     0      0 -0.7071068        0 -0.7071068      0
#> 2  2  7   1   0   0   0     0      0 -0.7071068        0 -0.7071068      0
#> 3  3  8   0   1   0   0    NA     NA         NA       NA         NA     NA
#> 4  4  9   0   0   1   0     1      1  0.7071068        1  0.7071068      1
#> 5  5 10   0   0   0   1     1      1  0.7071068        1  0.7071068      0
#>   f3na.c     f3naL3O.L  f3naL3O.Q  f3naL4O.L f3naL4O.Q  f3naL4O.C     f3naL5O.L
#> 1      0 -7.071068e-01  0.4082483 -0.6708204       0.5 -0.2236068 -6.324555e-01
#> 2      0 -7.071068e-01  0.4082483 -0.6708204       0.5 -0.2236068 -6.324555e-01
#> 3     NA            NA         NA         NA        NA         NA            NA
#> 4      0 -7.850462e-17 -0.8164966 -0.2236068      -0.5  0.6708204 -3.162278e-01
#> 5      1  7.071068e-01  0.4082483  0.2236068      -0.5 -0.6708204 -3.510833e-17
#>    f3naL5O.Q     f3naL5O.C  f3naL5O.4
#> 1  0.5345225 -3.162278e-01  0.1195229
#> 2  0.5345225 -3.162278e-01  0.1195229
#> 3         NA            NA         NA
#> 4 -0.2672612  6.324555e-01 -0.4780914
#> 5 -0.5345225  1.755417e-16  0.7171372