run machine learning models

This function compiles tools in caret packages to provide a machine learning (ML) interface where users only need to provide the name of the ML method and the rest are handled automatically.

Usage

ml.run(
  allDF = NULL,
  trainDF = NULL,
  testDF = NULL,
  preProc = c("center", "scale", "nzv"),
  classCol,
  posLabel = NULL,
  negLabel = NULL,
  cMethod,
  packName = NULL,
  trainInd = NULL,
  dummyEncoding = cMethod %in% c("xgbTree", "svmRadial"),
  prior.nzvCondX = NULL,
  prior.impute = c(NA, "missForest", "KNN"),
  oDir = ".",
  trainCtrlParas = NULL,
  trainParas = NULL
)

Arguments

allDF

NULL, data.frame; the data.frame of all the data where rows are observations and columns are features; training and testing will be sampled at the ratio of 75 vs 25.

trainDF

NULL, data.frame; the data.frame of the training data

testDF

NULL, data.frame; the data.frame of the testing data

preProc

c("center", "scale",'nzv'), char vector; preprocessing options; see caret::preProcess()

classCol

char; the column name of the class, i.e. the target variable

posLabel, negLabel

NULL, char; for classification only; the positive and negative labels for two-class classification

cMethod

char; the name of the ML method

packName

NULL, char; the name of the package containing the ML method. It's not needed if the cMethod is among those preset in method2package

trainInd

NULL, list of integers; the indices of the resampling iteration in CV; see index in caret::trainControl()

dummyEncoding

cMethod %in% c('xgbTree','svmRadial'),logical; is dummy variable encoding needed on the combined training and testing datasets before training? Although caret::train() handles dummy variable encoding automatically during CV and prediction on testing. It's possible that there are levels in the holdout or testing samples that are absent from the training data, causing errors.

prior.nzvCondX

NULL, char; remove NZV and conditionalX variables before training. This can be used only if caret::train is not used in the model building. Following are the available options:

NULL: do not remove before training

one of n, nx:

n: NZV

nx: NZV and ConditionalX if detected, removed variables are written to files nzv.csv and conditionalX.txt

prior.impute

c(NA,'missForest','KNN'), char; impute missing values before training. This can be used only if caret::train is not used in the model building. Following are the available options:

NA: do not impute before training

one of missForest, KNN: impute using missForest::missForest() or DMwR2::knnImputation()

oDir

NULL, char; output directory

trainCtrlParas

NULL, list; parameters for train control in caret::trainControl(). Here are its default values if method is not oob:

savePredictions = 'final',

if trainInd is provided, index = trainInd

otherwise method = "repeatedcv",repeats = 3,number = 10, which has the best bias-variance balance based on this post from Max Kuhn, the author of caret.

additional defaults for classification:

classProbs = TRUE,

summaryFunction = twoClassSummary or multiClassSummary depending on the class column

trainParas

NULL, list; parameters for train in caret::train(). Here are its default values:

tuneLength = 10

tuneLength is the number of levels for each tuning parameters that should be generated by train. If trainControl has the option search = "random", this is the maximum number of tuning parameter combinations that will be generated by the random search. if search = "grid" which is the default, this is the maximum number of levels of each tuning parameter. The actual number of levels used may be less than tuneLength. additional defaults for classification:

metric = "ROC" or 'Accuracy' depending on the class column

Value

a list of the following items

imp: data.frame, variable importance scores

trainStats: a named numerical vector containing AUC and various stats in the confusion table from the prediction on the training dataset.

testStats: similar, but from the testing dataset.

Details

The capabilities of this function include the following:

preprocessing of the whole dataset

dummy variable encoding

modification to the whole dataset before training, including

imputation using missForest and KNN

detection and removal of Near Zero Variance (NZV) and ConditionalX variables Note that these modifications apply only in situations where caret::train is not used in the model building; for example

✦ ML method is glm where no cross validation (CV) is needed due to the absence of hyper-parameters

✦ ML method is randomForest where the training is done using randomForest::tuneRF()

Otherwise, those modification should be specified in preProc to avoid information leakage during training. Note that missForest is not available in preProc.

training and testing

the output of various files of the training and testing stats:

nzv.csv: columns removed due to NZV and the details of each columns

train.all.stats.rds: stats of the prediction of training model on the training dataset

train.cv.pdf: curve of prediction performance during CV for upto four tuning parameters

train.model.rds

train.prediction.csv: prediction for each subject in training set

train.var.imp.pdf: plot of variable importance

train.var.imp.csv: importance score of each variable

test.all.stats.rds

test.prediction.csv

$cMethod.image.rdata: running environment for classification

train.confusion.table.txt

train.prediction.hist.pdf: prediction histogram

train.roc.pdf: ROC curve

train.MDS12.png, train.MDS13.png, train.MDS23.png: MDS plots of subjects if trained via randomForest

test.confusion.table.txt

test.prediction.hist.pdf

test.roc.pdf

test.MDS12.png, test.MDS13.png, test.MDS23.png

conditionalX.txt: columns removed due to conditionalX for regression

train.pred.cor.png: scatter plot of raw and predicted values

train.pred.residual.png: various diagnostic residual plots

test.pred.cor.png: scatter plot of raw and predicted values

test.pred.residual.png: plots of raw and standardized residuals

errors are handled to avoid crush.

Examples

# see https://blueskypie.github.io/ezML/articles/ezML-intro.html